Fixed to avoid related comparison of signed vs unsigned warnings in weapons.cpp
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@1964 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/GameWorld/weapon_definitions.h
@@ -236,7 +236,7 @@ static int16 weapon_ordering_array[]= {
 	_weapon_ball
 };
 
-#define NUMBER_OF_WEAPONS (sizeof(weapon_definitions)/sizeof(struct weapon_definition))
+#define NUMBER_OF_WEAPONS static_cast<int>(sizeof(weapon_definitions)/sizeof(struct weapon_definition))
 
 static struct weapon_definition weapon_definitions[]=
 {
